backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai-whisper
torch